Address feedback for styles, move them back under components directories.
