+++ features/base/styles/components/index.js
@@ -1,3 +1,3 @@
 export * from './_';
+export * from './ColorPalette';
 export * from './createStyleSheet';
-export * from './styles';
