+++ features/base/styles/components/native/shimStyles.js
@@ -1,8 +1,12 @@
 /**
  * Shim style properties to work correctly on native.
- *
+ * 
+ * Using this shimStyles method allows us to minimize the number of style
+ * declarations that need to be set or overridden for specific platforms.
+ * 
  * @param {Object} styles - A dictionary of named style definitions.
- * @returns {Object} */
+ * @returns {Object}
+ */
 export function shimStyles(styles) {
     return styles;
 }
