--- features/base/styles/components/styles/index.js
@@ -1,1 +1,0 @@
-export * from './ColorPalette';
