+++ features/conference/index.js
@@ -1,2 +1,2 @@
 export * from './components';
-export { default as styles } from './styles';
+
