+++ features/filmStrip/components/native/VideoThumbnailContainer.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { TouchableHighlight, View } from 'react-native';
 
-import styles from '../../styles';
+import { styles } from './styles';
 
 /**
  * Web version of Video Thumbnail Container component.
