+++ features/filmStrip/components/web/AudioMutedIndicator.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Icon from 'react-fontawesome';
 
-import styles from '../../styles';
+import { styles } from './styles';
 
 /**
  * Thumbnail badge for displaying the audio mute status of a participant.
