+++ features/filmStrip/components/web/DominantSpeakerIndicator.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Icon from 'react-fontawesome';
 
-import styles from '../../styles';
+import { styles } from './styles';
 
 /**
  * Thumbnail badge showing that the participant is the dominant speaker in
