+++ features/filmStrip/components/web/VideoMutedIndicator.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Icon from 'react-fontawesome';
 
-import styles from '../../styles';
+import { styles } from './styles';
 
 /**
  * Thumbnail badge for displaying the video mute status of a participant.
