--- features/filmStrip/styles/_.ios.js
@@ -1,1 +1,0 @@
-export { default as styles } from './native/Styles';
