--- features/filmStrip/styles/_.web.js
@@ -1,1 +1,0 @@
-export { default as styles } from './web/Styles';
