--- features/filmStrip/styles/index.js
@@ -1,3 +1,0 @@
-import Styles from './Styles';
-
-export default Styles;
