+++ features/toolbar/components/styles/Styles.js
@@ -1,10 +1,10 @@
-import { ColorPalette, createStyleSheet } from '../../base/styles';
+import { ColorPalette, createStyleSheet } from '../../../base/styles';
 
 /**
  * The toolbar related styles.
  * TODO Make styles more generic and reusable. Use ColorPalette for all colors.
  */
-const styles = createStyleSheet({  
+export const styles = createStyleSheet({  
     /**
      * The toolbar button icon style.
      */
@@ -44,5 +44,3 @@ const styles = createStyleSheet({
         right: 0
     }
 });
-
-export default styles;
