+++ features/toolbar/components/web/ToolbarContainer.js
@@ -3,7 +3,7 @@ import Icon from 'react-fontawesome';
 
 import { ColorPalette } from '../../../base/styles';
 
-import styles from '../../styles';
+import { styles } from '../styles';
 
 /**
  * The web container rendering the toolbar.
