+++ features/welcome/components/native/WelcomePageContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, TextInput, TouchableHighlight, View } from 'react-native';
 
-import styles from '../../styles';
+import { styles } from '../styles';
 
 /**
  * The native container rendering the welcome page.
