+++ features/welcome/components/styles/Styles.js
@@ -1,11 +1,11 @@
-import { ColorPalette, createStyleSheet } from '../../base/styles';
+import { ColorPalette, createStyleSheet } from '../../../base/styles';
 
 /**
  * The welcome page style.
  * TODO: Make styles more generic and reusable. Use color palette for all
  * colors.
  */
-var styles = createStyleSheet({
+export const styles = createStyleSheet({
     /**
      * Join button text style.
      */
@@ -75,5 +75,3 @@ var styles = createStyleSheet({
         textAlign: 'center'
     }
 });
-
-export default styles;
