+++ features/welcome/components/styles/index.js
@@ -1,0 +1,1 @@
+export * from './Styles';
